requirements.txt
# Backend
fastapi==0.109.2
uvicorn[standard]==0.27.1
httpx==0.26.0
python-dotenv==1.0.1
openai==1.12.0

# CORS
python-multipart==0.0.9
